requirements.txt
astroid==2.3.3
bleach==3.1.0
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
cycler==0.10.0
docutils==0.15.2
freetype-py==2.1.0.post1
idna==2.8
isort==4.3.21
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
matplot==0.1.9
matplotlib==3.1.1
mccabe==0.6.1
mysql-connector-python==8.0.17
numpy==1.17.3
pandas==0.25.3
pkginfo==1.5.0.1
pygame==1.9.6
Pygments==2.4.2
pylint==2.4.3
pyloco==0.0.139
pyparsing==2.4.4
python-dateutil==2.8.1
pytz==2019.3
readme-renderer==24.0
requests==2.22.0
requests-toolbelt==0.9.1
SimpleWebSocketServer==0.1.1
six==1.13.0
tqdm==4.37.0
twine==2.0.0
typed-ast==1.4.0
typing==3.7.4.1
urllib3==1.25.6
ushlex==0.99.1
virtualenv==16.7.7
webencodings==0.5.1
websocket-client==0.56.0
wrapt==1.11.2
